ci: add benchmarks action, extract folder as a module
