+++ benchmarks/src/main/java/com/artipie/conda/CondaRepodataRemoveBench.java
@@ -2,9 +2,8 @@
  * The MIT License (MIT) Copyright (c) 2020-2021 artipie.com
  * https://github.com/artipie/conda-adapter/LICENSE
  */
-package com.artipie.conda.benchmarks;
+package com.artipie.conda;
 
-import com.artipie.conda.CondaRepodata;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
